requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.0
google-auth==2.39.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
motor==3.7.0
multidict==6.4.3
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyjwt==2.10.1
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pyyaml==6.0.2
realtime==2.4.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
